כרטיסי ההגדרות: ריכוז הקוד בווידג'ט אחד, וחלוקת ווידג'ט כפתורים בין הגדרות לשאר התוכנה
