+++ lib/settings/tabs/system_settings_tab.dart
@@ -36,8 +36,7 @@ import 'package:otzaria/data/repository/data_repository.dart';
 import 'package:otzaria/widgets/widgets_exports.dart';
 import 'package:otzaria/widgets/dialogs/error_report_sender_email_dialog.dart';
 import 'package:otzaria/widgets/text/rtl_text_field.dart';
-import 'package:otzaria/settings/widgets/settings_card.dart';
-import 'package:otzaria/settings/widgets/settings_action_tile.dart';
+import 'package:otzaria/settings/widgets/settings_widgets_exports.dart';
 import 'package:otzaria/theme/theme_exports.dart';
 import 'package:otzaria/text_book/view/error_report_dialog.dart';
 import 'package:otzaria/tour/bloc/tour_cubit.dart';
