+++ lib/settings/widgets/settings_widgets_exports.dart
@@ -2,3 +2,4 @@ export 'settings_card.dart';
 export 'switch_settings_tile.dart';
 export 'settings_action_tile.dart';
 export 'dropdown_settings_tile.dart';
+export 'segmented_settings_tile.dart';
