+++ test/widgets/segmented_settings_tile_test.dart
@@ -2,6 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:fluentui_system_icons/fluentui_system_icons.dart';
+import 'package:otzaria/settings/widgets/segmented_settings_tile.dart';
 import 'package:otzaria/widgets/inputs/segmented_button_tile.dart';
 
 void main() {
